Move `Coerce` impls into `coerce` mod.
