+++ eval/src/value.rs
@@ -1,4 +1,4 @@
-use crate::{Attrs, Coerce, Func, List, Object};
+use crate::{List, Object};
 use derive_more::From;
 use std::fmt;
 
@@ -8,40 +8,11 @@ pub enum Value {
     List(List),
     Object(Object),
 }
-use Value::*;
-
-impl Coerce for f64 {
-    fn coerce_from_value(v: &Value) -> Option<&f64> {
-        match v {
-            Num(x) => Some(x),
-            _ => None,
-        }
-    }
-}
-
-impl Coerce for Object {
-    fn coerce_from_value(v: &Value) -> Option<&Object> {
-        match v {
-            Object(x) => Some(x),
-            _ => None,
-        }
-    }
-}
-
-impl Coerce for Func {
-    fn coerce_from_value(v: &Value) -> Option<&Func> {
-        Object::coerce_from_value(v).and_then(|obj| obj.func())
-    }
-}
-
-impl Coerce for Attrs {
-    fn coerce_from_value(v: &Value) -> Option<&Attrs> {
-        Object::coerce_from_value(v).map(|obj| obj.attrs())
-    }
-}
 
 impl fmt::Display for Value {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        use Value::*;
+
         match self {
             Num(x) => x.fmt(f),
             List(x) => x.fmt(f),
